Login and registration

+++ routes/web.php
@@ -1,16 +1,25 @@
 <?php
 
-use App\Http\Controllers\ListingController;
+use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Models\Listing;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\ListingController;
 
-
+// Job Listing
 Route::get('/', [ListingController::class, 'showAll']);
-Route::get('/listings/create', [ListingController::class, 'create']);
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
-Route::post('/listings', [ListingController::class, 'store']);
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
-Route::delete('/listings/{listing}', [ListingController::class, 'delete']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
+Route::delete('/listings/{listing}', [ListingController::class, 'delete'])->middleware('auth');
 Route::get('/listings/{listing}', [ListingController::class, 'showOne']);
+
+// User
+Route::get('/register', [UserController::class, 'register'])->middleware('guest');
+Route::post('/users', [UserController::class, 'store']);
+Route::post('/users/login', [UserController::class, 'loginUser']);
+
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+Route::post('/logout', [UserController::class, 'logout']);
